+++ projects/pip-webui2-behaviors/src/lib/shortcuts/shared/ShortcutOptions.ts
@@ -1,0 +1,13 @@
+export enum KeyboardEventType {
+    Keydown = 'keydown',
+    Keyup = 'keyup',
+    Keypress = 'keypress'
+}
+
+export class ShortcutOptions {
+    Type: KeyboardEventType;
+    Propagate: boolean;
+    DisableInInput: boolean;
+    Target: any; // Event target object
+    Keycode: number;
+}