+++ projects/pip-webui2-behaviors/src/lib/shortcuts/shared/ShortcutsConfig.ts
@@ -1,0 +1,12 @@
+import { ShortcutItem } from './ShortcutItem';
+import { ShortcutOptions } from './ShortcutOptions';
+
+export class ShortcutsConfig {
+    /// Global Shortcuts
+    public globalShortcuts: ShortcutItem[] = [];
+    /// Local shortcuts
+    public localShortcuts: ShortcutItem[] = [];
+    /// Default options
+    public defaultOptions: ShortcutOptions = null;
+}
+