+++ projects/pip-webui2-behaviors/src/lib/shortcuts/shortcut.module.ts
@@ -1,0 +1,20 @@
+import { NgModule, ModuleWithProviders } from '@angular/core';
+
+import { PipShortcutsService as PipShortcutsService } from './shared/shortcut.service';
+import { PipShortcutsBindingService, shortcutsDefaultProvider } from './shared/shortcut.binding.service';
+
+@NgModule({
+    declarations: [],
+})
+export class PipShortcutsModule {
+    static forRoot(): ModuleWithProviders {
+        return {
+            ngModule: PipShortcutsModule,
+            providers: [
+                PipShortcutsBindingService,
+                PipShortcutsService,
+                shortcutsDefaultProvider
+            ]
+        };
+    }
+}